Reimplemented database tests.
